Hw 15

+++ src/features/blogs/posts/application/use-cases/delete-post.use-case.ts
@@ -14,6 +14,16 @@ export class DeletePostUseCase implements ICommandHandler<DeletePostCommand> {
   constructor(private readonly postsRepository: PostsRepository) {}
 
   async execute(command: DeletePostCommand) {
+    if (!command.id) {
+      throw new NotFoundException('id not found');
+    }
+
+    const isExisted = await this.postsRepository.isExistedPost(command.id);
+
+    if (!isExisted) {
+      throw new NotFoundException();
+    }
+
     const result = this.postsRepository.deletePost(command.id);
 
     if (!result) {
